services/requirements.txt
nltk
mysql-connector-python